pyproject.toml
[tool.poetry]
name = "argilla-streamlit"
version = "0.1.0"
description = "👑 Streamlit for extended UI functionalities for Argilla."
authors = ["david <david.m.berenstein@gmail.com>"]
readme = "README.md"
packages = [{include = "argilla_streamlit"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
pandas = "^1.5.3"
spacy = "^3.5.0"
argilla = {extras = ["listeners"], version = "^1.3.1"}
streamlit = "1.17"
watchdog = "^2.2.1"
text-highlighter = "^0.0.11"
datasets = "^2.9.0"
huggingface-hub = "^0.12.1"
streamlit-tags = "^1.2.8"
streamlit-datalist = "^0.0.3"
openpyxl = "^3.1.1"
XlsxWriter = "^3.0.8"
streamlit-plotly-events = "^0.0.6"
plotly-express = "^0.4.1"
umap-learn = "^0.5.3"
bokeh = "2.4.3"
streamlit-bokeh-events = "^0.1.2"
pysbd = "^0.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
